decouple on download, detail click
