+++ src/components/result/SelectedListCard.tsx
@@ -7,26 +7,17 @@ import { OGCCollection } from "../common/store/OGCCollectionDefinitions";
 
 interface SelectedListCardProps {
   content: OGCCollection;
-  onDownload: (uuid: string, tab: string, section?: string) => void;
-  onLink: (uuid: string, tab: string, section?: string) => void;
-  onDetail: (uuid: string) => void;
   onClickCard: (uuid: string) => void;
 }
 
 const SelectedListCard: React.FC<SelectedListCardProps> = ({
   content,
-  onDownload,
-  onLink,
-  onDetail,
   onClickCard,
 }) => {
   return (
     <Box height={LIST_CARD_HEIGHT - LIST_CARD_GAP * 2} mb={gap.lg}>
       <ListResultCard
         content={content}
-        onDownload={onDownload}
-        onLink={onLink}
-        onDetail={onDetail}
         onClickCard={onClickCard}
         isSelectedDataset
       />
