+++ src/pages/search-page/subpages/ResultSection.tsx
@@ -35,8 +35,6 @@ const ResultSection: FC<ResultSectionProps> = ({
   onChangeSorting,
   onClickCard,
   onDetail,
-  onDownload,
-  onLink,
   isLoading,
 }) => {
   // Get contents from redux
@@ -107,8 +105,6 @@ const ResultSection: FC<ResultSectionProps> = ({
           <ResultCards
             layout={currentLayout}
             contents={reduxContents}
-            onDownload={onDownload}
-            onLink={onLink}
             onDetail={onDetail}
             onClickCard={onClickCard}
             onFetchMore={fetchMore}
